+++ src/controllers/user.ts
@@ -1,9 +1,11 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 // import firestore from '@google-cloud/firestore'
 import { CustomNodeJSGlobal } from '../custom/global'
+import { List } from './list'
 import { DtoUser } from '../dto-interfaces/user.dto'
+import { DtoList } from '../dto-interfaces/list.dto'
 import { IUser } from '../interfaces/user'
-import { EFU, MFU } from './utils/user.messages'
+import { CFU, EFU, MFU } from './utils/user.messages'
 import { encryptMessage, generatePassword } from '../utils/security'
 import { mail } from '../utils/mailer'
 import { MFME } from '../utils/messages'
@@ -29,8 +31,15 @@ class User {
     this._result = []
   }
 
-  public process (type: string): Promise<string> | undefined {
+  public process (
+    type   : string,
+    listId?: string,
+  ): Promise<string> | undefined {
     switch (type) {
+      case 'enrollStudent':
+        return this._enroll('student', listId as string)
+      case 'enrollTeacher':
+        return this._enroll('teacher', listId as string)
       case 'notifyStudent':
         return this._notify('student')
       case 'notifyTeacher':
@@ -41,6 +50,79 @@ class User {
         return this._verifyUser('teacher')
       default:
         return undefined
+    }
+  }
+
+  private async _enroll (
+    condition: string,
+    listId   : string,
+  ): Promise<string> {
+    const document = this._args?.documentType === '0'? 'documentNumber' : 'UNICode'
+    let user: any
+
+    try {
+      if (condition === 'teacher')
+        user = await this._teachersRef
+          .where(document, '==', `${this._args?.documentNumber}`)
+          .get()
+      else
+        user = await this._studentsRef
+          .where(document, '==', `${this._args?.documentNumber}`)
+          .get()
+
+      const userData = user.docs.map((doc: any) => {
+        return {
+          ...doc.data(),
+          id: doc.id
+        }
+      })[0]
+
+      if ('postulating' in userData && userData.postulating)
+        throw condition === 'teacher'
+          ? new Error(`${CFU.article}${CFU.teacher}${EFU.errorEnrolling1}`)
+          : new Error(`${CFU.article}${CFU.student}${EFU.errorEnrolling1}`)
+
+      if ('password' in userData && userData.password)
+        throw condition === 'teacher'
+          ? new Error(`${CFU.article}${CFU.teacher}${EFU.errorEnrolling2}`)
+          : new Error(`${CFU.article}${CFU.student}${EFU.errorEnrolling2}`)
+
+      if (condition === 'teacher')
+        await this._teachersRef
+          .doc(userData.id as string)
+          .update({
+            list       : listId,
+            postulating: true
+          })
+      else
+        await this._studentsRef
+          .doc(userData.id as string)
+          .update({
+            list       : listId,
+            postulating: true
+          })
+
+      const l = new List({
+        id   : listId,
+        owner: userData.id as string,
+        type : condition
+      } as DtoList)
+
+      await l.enroll(userData.id as string, condition)
+
+      return encryptMessage(JSON.stringify(userData), KEY_JSON)
+    } catch (error) {
+      console.error(error)
+
+      if (
+        error.message === `${CFU.article}${CFU.teacher}${EFU.errorEnrolling1}` ||
+        error.message === `${CFU.article}${CFU.student}${EFU.errorEnrolling1}`
+      )
+        throw error
+
+        throw condition === 'teacher'
+          ? new Error(`${EFU.errorEnrolling}${CFU.pTeacher}`)
+          : new Error(`${EFU.errorEnrolling}${CFU.pStudent}`)
     }
   }
 
@@ -70,20 +152,21 @@ class User {
       // Updating password
       if (condition === 'teacher')
         await this._teachersRef
-          .doc(this._args.id as string)
+          .doc(data.id as string)
           .update({ password: newPassword.ePassword })
       else
         await this._studentsRef
-          .doc(this._args.id as string)
+          .doc(data.id as string)
           .update({ password: newPassword.ePassword })
 
       return encryptMessage(MFU.updateAndNotifySuccess, KEY_JSON)
     } catch (error) {
       console.log(error)
+
       if (error.message === MFME.generic)
         throw error
 
-      throw new Error(EFU.errorUpdating)
+      throw new Error(EFU.errorNotifying)
     }
   }
 
@@ -114,14 +197,17 @@ class User {
 
       return encryptMessage(JSON.stringify(this._result[0]), KEY_JSON)
     } catch (error) {
+      console.error(error)
+
       if (
         error.message === EFU.teacherNotFound ||
         error.message === EFU.studentNotFound
       )
         throw error
 
-      console.error(error)
-      throw new Error(`There was a error verifying the ${condition}`)
+      throw condition === 'teacher'
+        ? new Error(`${EFU.errorVerifying}${CFU.pTeacher}`)
+        : new Error(`${EFU.errorVerifying}${CFU.pStudent}`)
     }
   }
 }