+++ src/controllers/utils/list.messages.ts
@@ -1,0 +1,13 @@
+enum ErrorMessagesForLists {
+  errorCreating = 'Hubo un error creando la lista.',
+  errorEnrolling = 'Hubo un problema inscribiendo al '
+}
+
+enum MessagesForLists {
+  success = 'Su lista fue creada exitosamente.'
+}
+
+export {
+  ErrorMessagesForLists as EFL,
+  MessagesForLists as MFL
+}