+++ src/dto-interfaces/list.dto.ts
@@ -1,0 +1,5 @@
+export interface DtoList {
+  id?   : string
+  owner?: string
+  type  : string
+}