+++ src/dto-interfaces/user.dto.ts
@@ -2,5 +2,4 @@ export interface DtoUser {
   documentNumber?: string
   documentType?  : string
   id?            : string
-  password?      : string
 }