+++ src/interfaces/list.ts
@@ -1,0 +1,6 @@
+export interface IList {
+  applicants?: string[]
+  id         : string
+  owner?     : string
+  type?      : string
+}