+++ src/routes/index.ts
@@ -1,7 +1,9 @@
 import { Home } from './home'
+import { List } from './list'
 import { User } from './user'
 
 export {
   Home,
+  List,
   User
 }